Saved min_fluor data using guidata
The settings approach wasn't working for me. It was creating a new
variable with new fields. So I looked up the documentation and tried
using guidata to save my changes to handles. It worked! I can now see
updated images.
I still have a few questions:
A) should I put an upper-bound on MIN_FLUOR? This doesn't seem
necessary.
B) should I allow decimal values? This seems useful since most of the
fluors in the sample images are in the range 1-2 and so we don't get
much granularity by using an integer threshold.